Move diagnosticContext from rootRef to boskInfo.
It was misplaced on RootReference.

I believe I added it there so that a component that has access to only a
Reference could get its hands on the diagnostic context. However, it's
unclear whether this is an important capability: under what
circumstances would there be multiple DiagnosticContexts in play, and
the right one to pick would vary dynamically on a reference-by-reference
basis? Really, whatever component needs the DiagnosticContext should be
designed to accept the right context at creation time and use it.

It's possible I might some day remember why I thought I needed the
DiagnosticContext to come from a Reference, but in the mean time, it
just looked funny.