Virtual threads to isolate ThreadLocals in hooks.
There's something kind of funny about going to great lengths to use our
Reactive Breadth First Search algorithm to achieve the hook ordering we
want without using a background thread, and then right at the last
moment, spawning a thread to run each hook anyway.

I'm leaving it this way for the time being because it has certain
desirable properties:

- The hook queue maintains the desired order. In contrast, if we simply
  submitted the hooks to an executor, all bets are off.
- Only one hook runs at a time. There's a happens-before relationship
  between all the hooks.

In effect, since we're running completely synchronously with the virtual
thread, using it as a kind of structured concurrency sans parallelism,
we're actually just using the virtual thread mechansim as a utility to
isolate ThreadLocals and nothing else.

Anyway, this is a step closer to the semantics we want, and we can
refine the implementation later if we can devise a simpler way to
achieve all the desired properties.