MongoDriverSettings.timescaleMS.
Combine recoveryPollingMS and flushTimeoutMS into a single "timescale" setting.

Tuning a multitude of individual timeouts is not a reasonable thing to
ask users to do. Just ask them about how responsive they want the system
to be, and then we can take the responsibility of balancing the actual
individual timeout settings in a way that makes sense given how all the
parts of this rather complex driver work together.

This is something I already did in the SQL driver. I might end up
copying over the patienceFactor idea too.