Restore CurriedFields in ClassBuilder.
Now that the CallSite table is immortal, everything added to that is a
memory leak, so we need to be judicious. Tests with many Bosk objects
having large state trees can run out of memory, which is pretty
annoying.

The downside of curried fields is that final instance fields are not
_really_ final, so they're generally slower than static finals and
ConstantCallSites, but at the moment, OOMs caused by leaks seem to be a
bigger problem (though unlikely in production, where there's only
typically a single, long-lived Bosk object).