Stop using InjectFrom for constructor parameters.
This already works well with @ParameterizedClass, so let's encourage
people to use that. My goal is to give class-level annotations to inject
method parameters, but class-level parameters (for constructors or
@Parameter fields) are already well served by built-in JUnit features.

The two can cooperate nicely: if you define a ParameterInjector, you can
always use it for constructor parameters by specifying a MethodSource
that instantiates your injector and calls values(). There are several
examples of that in this commit.