Rename Bosk.Info.
It was called UnderConstruction because the intent was that this
represented a view of the bosk from before the constructor had finished.
However, it's perfectly valid for driver factories to hold on to a
reference to the BoskInfo object, and as a result, it was confusing when
debugging to encounter an object called "UnderConstruction" for a bosk
that was already fully constructed.

Let's change it to a more noncomittal name that will not be misleading
when folks are debugging their applications.