Simplify TransactionalCollection.
This used to inherit the MongoDB client library's Collection interface,
and was intended as a drop-in replacement. However, that means every
time that interface changes, TransactionalCollection breaks, and as a
corollary, it means bosk-mongo is compatible with only a narrow range of
client library versions.

There's no real reason this should implement the client library's
interface. Instead, let's treat it as our own encapsulation of the
(rather cumbersome and error prone) Collection interface for our own needs.