In FAIL_FAST, try opening change stream cursor.
I expected it to be a common mistake for people to forget to use a
replica set. In FAIL_FAST mode, it seems like that ought to reliably
throw an exception before the Bosk constructor returns.

Still won't help much as long as the default is DISCONNECT instead of
FAIL_FAST though.