Fix: Remove maxAwaitTime from openCursor.
I think this was needed before ChangeReceiver.interrupt so that the
ChangeReceiver would notice the loss of connectivity, and subsequently
reconnect when connectivity is restored.  Now that we have
ChangeReceiver.interrupt, the ChangeReceiver doesn't need a timeout to
notice a loss of connectivity anymore.

The ChangeReceiver also uses the Mongo client's heartbeat system to
notice a connectivity loss even if there is no bosk driver activity.